% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdgt_pal.R
\docType{data}
\name{spdgt_pal}
\alias{spdgt_pal}
\title{SpeedGoat color palette}
\format{
A character vector of length 8
}
\source{
Developed by SpeedGoat
}
\usage{
spdgt_pal
}
\description{
A vector of hex codes stored as data.
}
\keyword{datasets}
